文章分类更新优化：增加是否变更判断

+++ app/controllers/backend/PostController.php
@@ -109,21 +109,25 @@ class PostController extends BaseController {
 
         $post->post_title = Input::get('title');
         $post->post_content = Input::get('content');
+        $post->save();
 
         //因为打算改成文章有多个分类，所以这样的操作是合理的
-        $post->termRelation()->delete();
         $categories = Input::get('category');
         is_array($categories) || $categories = array($categories);
-        $termRelMultiData = array_map(function($categoryId) use ($post){
-            return array(
-                    'object_id'   => $post->id,
-                    'category_id' => $categoryId
-                   );
-        }, $categories);
-        //一次性写入多条
-        //请参考：http://www.golaravel.com/docs/4.1/queries/#inserts
-        DB::table('term_relationships')->insert($termRelMultiData);
-        $post->save();
+
+        //如果不同才存
+        if (Input::get('old_category', '') != join(',', $categories)) {
+            $post->termRelation()->delete();
+            $termRelMultiData = array_map(function($categoryId) use ($post){
+                return array(
+                        'object_id'   => $post->id,
+                        'category_id' => $categoryId
+                       );
+            }, $categories);
+            //一次性写入多条
+            //请参考：http://www.golaravel.com/docs/4.1/queries/#inserts
+            DB::table('term_relationships')->insert($termRelMultiData);    
+        }
 
         return Redirect::back()->withMessage('更新成功！');
     }
